Moved everything into a modular crate
