+++ Cargo.toml
@@ -9,6 +9,7 @@ members = [
     "livekit-ffi",
     "livekit-uniffi",
     "livekit-datatrack",
+    "livekit-token-source",
     "livekit-ffi-node-bindings",
     "livekit-runtime",
     "livekit-wakeword",
@@ -53,6 +54,7 @@ livekit = { version = "0.7.53", path = "livekit" }
 livekit-api = { version = "0.5.6", path = "livekit-api" }
 livekit-ffi = { version = "0.12.71", path = "livekit-ffi" }
 livekit-datatrack = { version = "0.1.11", path = "livekit-datatrack" }
+livekit-token-source = { version = "0.1.0", path = "livekit-token-source" }
 livekit-common = { version = "0.1.0", path = "livekit-common" }
 livekit-data-stream = { version = "0.1.0", path = "livekit-data-stream" }
 livekit-protocol = { version = "0.7.10", path = "livekit-protocol" }
