+++ livekit-token-source/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "livekit-token-source"
+description = "Token sources for the LiveKit Rust SDK"
+version = "0.1.0"
+license.workspace = true
+edition.workspace = true
+repository.workspace = true
+readme = "README.md"
+
+[dependencies]
+thiserror = { workspace = true }
