+++ livekit-token-source/src/response.rs
@@ -1,0 +1,14 @@
+use crate::error::TokenSourceError;
+
+pub struct TokenSourceResponse {
+    pub server_url: String,
+    pub participant_token: String,
+}
+
+impl TokenSourceResponse {
+    pub fn new(server_url: String, participant_token: String) -> TokenSourceResponse {
+        TokenSourceResponse{server_url: server_url, participant_token: participant_token}
+    }
+}
+
+pub type TokenSourceResult<T> = Result<T, TokenSourceError>;
